Restructure Skootrs and simplify abstractions
This restructures Skootrs to be focused around a handful of
directories.

bin/ This is where the CLI code lives

model/ This is where the data important to GUAC lives.

model/skootrs This is where skootrs specific data models go. Other
external data models go in other directories.

server/ This is where the rest and any other web app code goes.
As much as is possible this should contain no real logic outside
of what is required to manage a rest server. The logic should be
handled in services.

service/ This is where the majority of logic is managed. This
contains the mechanisms to create and manage repos, files, etc.