Convert skootrs to cargo workspace
This converts skootrs from one monolithic crate to a workspace
with 4 subcrates to speed up compilation and make vscode less
laggy.

skootrs-bin - The crate for the CLI
skootrs-lib - The core crate that handles most of the
functionality and data models
skootrs-rest - The crate for the rest/openapi server
skootrs-statestore - The crate for managing the persistence
of state data.