Refactor CLI
This standardizes the CLI on noun-verb commands. This also begins
to structure the command line helpers to be more generic and
take in things like config and only do the interactive prompt if
params from a file, stdin, url, or pipe in yaml or json format
is not passed.

Right now only project create functionality has been completely
refactored in this way, but the other commands shouldn't be
much longer